Fix tests to reflect the move of modules to exe package
